chore: consolidate cowswap utils (#8331)
