+++ packages/swapper/src/cowswap-utils/index.ts
@@ -1,32 +1,34 @@
+import type { LatestAppDataDocVersion } from '@cowprotocol/app-data'
+import { MetadataApi, stringifyDeterministic } from '@cowprotocol/app-data'
+import type { OrderClass, OrderClass1 } from '@cowprotocol/app-data/dist/generatedTypes/v1.3.0'
 import type { ChainId } from '@shapeshiftoss/caip'
-import { fromChainId } from '@shapeshiftoss/caip'
+import { ASSET_NAMESPACE, fromChainId, toAssetId } from '@shapeshiftoss/caip'
 import type { EvmChainAdapter, SignTypedDataInput } from '@shapeshiftoss/chain-adapters'
 import { toAddressNList } from '@shapeshiftoss/chain-adapters'
 import type { ETHSignTypedData, HDWallet } from '@shapeshiftoss/hdwallet-core'
 import type { AccountMetadata, TypedDataTypes, UnsignedOrderCreation } from '@shapeshiftoss/types'
 import { CowNetwork, KnownChainIds, TypedDataPrimaryType } from '@shapeshiftoss/types'
+import {
+  bnOrZero,
+  convertDecimalPercentageToBasisPoints,
+  getNativeFeeAssetReference,
+} from '@shapeshiftoss/utils'
 import type { TypedData } from 'eip-712'
 import type { TypedDataDomain } from 'ethers'
 import { ethers } from 'ethers'
+import type { Address } from 'viem'
+import { keccak256, stringToBytes } from 'viem'
 
-import { COW_SWAP_SETTLEMENT_ADDRESS } from '../swappers/CowSwapper'
+import type { AffiliateAppDataFragment } from '../swappers/CowSwapper'
+import { getTreasuryAddressFromChainId } from '../swappers/utils/helpers/helpers'
+import {
+  CANCELLATIONS_TYPE_FIELDS,
+  COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS,
+  COW_SWAP_SETTLEMENT_ADDRESS,
+  ORDER_TYPE_FIELDS,
+} from './constants'
 
-export const ORDER_TYPE_FIELDS = [
-  { name: 'sellToken', type: 'address' },
-  { name: 'buyToken', type: 'address' },
-  { name: 'receiver', type: 'address' },
-  { name: 'sellAmount', type: 'uint256' },
-  { name: 'buyAmount', type: 'uint256' },
-  { name: 'validTo', type: 'uint32' },
-  { name: 'appData', type: 'bytes32' },
-  { name: 'feeAmount', type: 'uint256' },
-  { name: 'kind', type: 'string' },
-  { name: 'partiallyFillable', type: 'bool' },
-  { name: 'sellTokenBalance', type: 'string' },
-  { name: 'buyTokenBalance', type: 'string' },
-]
-
-export const CANCELLATIONS_TYPE_FIELDS = [{ name: 'orderUids', type: 'bytes[]' }]
+export * from './constants'
 
 export const getCowNetwork = (chainId: ChainId): CowNetwork | undefined => {
   switch (chainId) {
@@ -156,4 +158,80 @@ export const signCowOrderCancellation = async (
   const signature = await signMessage(typedData)
 
   return signature
+}
+
+export const cowSwapTokenToAssetId = (chainId: ChainId, cowSwapToken: Address) => {
+  const { chainNamespace, chainReference } = fromChainId(chainId)
+  return cowSwapToken.toLowerCase() === COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS.toLowerCase()
+    ? toAssetId({
+        chainId,
+        assetNamespace: 'slip44',
+        assetReference: getNativeFeeAssetReference(chainNamespace, chainReference),
+      })
+    : toAssetId({
+        chainId,
+        assetNamespace: ASSET_NAMESPACE.erc20,
+        assetReference: cowSwapToken,
+      })
+}
+
+export const getAffiliateAppDataFragmentByChainId = ({
+  affiliateBps,
+  chainId,
+}: {
+  affiliateBps: string
+  chainId: ChainId
+}): AffiliateAppDataFragment => {
+  const hasAffiliateFee = bnOrZero(affiliateBps).gt(0)
+  if (!hasAffiliateFee) return {}
+
+  return {
+    partnerFee: {
+      bps: Number(affiliateBps),
+      recipient: getTreasuryAddressFromChainId(chainId),
+    },
+  }
+}
+
+type AppDataInfo = {
+  doc: LatestAppDataDocVersion
+  fullAppData: string
+  appDataKeccak256: string
+  env?: string
+}
+
+const generateAppDataFromDoc = async (
+  doc: LatestAppDataDocVersion,
+): Promise<Pick<AppDataInfo, 'fullAppData' | 'appDataKeccak256'>> => {
+  const appData = await stringifyDeterministic(doc)
+  const appDataKeccak256 = keccak256(stringToBytes(appData))
+
+  return { fullAppData: appData, appDataKeccak256 }
+}
+
+const metadataApi = new MetadataApi()
+
+// See https://api.cow.fi/docs/#/default/post_api_v1_quote / https://github.com/cowprotocol/app-data
+export const getFullAppData = async (
+  slippageTolerancePercentage: string,
+  affiliateAppDataFragment: AffiliateAppDataFragment,
+  orderClass1: OrderClass1,
+) => {
+  const APP_CODE = 'shapeshift'
+  const orderClass: OrderClass = { orderClass: orderClass1 }
+  const quote = {
+    slippageBips: convertDecimalPercentageToBasisPoints(slippageTolerancePercentage).toNumber(),
+  }
+
+  const appDataDoc = await metadataApi.generateAppDataDoc({
+    appCode: APP_CODE,
+    metadata: {
+      quote,
+      orderClass,
+      ...affiliateAppDataFragment,
+    },
+  })
+
+  const { fullAppData, appDataKeccak256 } = await generateAppDataFromDoc(appDataDoc)
+  return { appDataHash: appDataKeccak256, appData: fullAppData }
 }
