+++ packages/swapper/src/swappers/CowSwapper/filterBuyAssetsBySellAssetId/filterBuyAssetsBySellAssetId.ts
@@ -1,9 +1,9 @@
 import type { AssetId } from '@shapeshiftoss/caip'
 
+import { SUPPORTED_CHAIN_IDS } from '../../../cowswap-utils/constants'
 import type { BuyAssetBySellIdInput } from '../../../types'
 import { isNativeEvmAsset } from '../../utils/helpers/helpers'
 import { COWSWAP_UNSUPPORTED_ASSETS } from '../utils/blacklist'
-import { SUPPORTED_CHAIN_IDS } from '../utils/constants'
 
 export const filterBuyAssetsBySellAssetId = ({
   assets,
