+++ packages/swapper/src/swappers/CowSwapper/index.ts
@@ -1,2 +1,1 @@
 export * from './types'
-export * from './utils'
