--- packages/swapper/src/swappers/CowSwapper/utils/index.ts
@@ -1,1 +1,0 @@
-export { COW_SWAP_SETTLEMENT_ADDRESS } from './constants'
