+++ src/state/apis/limit-orders/limitOrderApi.ts
@@ -4,15 +4,13 @@ import { fromAccountId, fromAssetId } from '@shapeshiftoss/caip'
 import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import {
   assertGetCowNetwork,
+  COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS,
+  getAffiliateAppDataFragmentByChainId,
+  getFullAppData,
   signCowMessage,
   signCowOrder,
   signCowOrderCancellation,
 } from '@shapeshiftoss/swapper'
-import { COW_SWAP_NATIVE_ASSET_MARKER_ADDRESS } from '@shapeshiftoss/swapper/dist/swappers/CowSwapper/utils/constants'
-import {
-  getAffiliateAppDataFragmentByChainId,
-  getFullAppData,
-} from '@shapeshiftoss/swapper/dist/swappers/CowSwapper/utils/helpers/helpers'
 import { isNativeEvmAsset } from '@shapeshiftoss/swapper/dist/swappers/utils/helpers/helpers'
 import type {
   CowSwapError,
